third_party/WebKit/LayoutTests/fast/workers/worker-base64.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
<link rel="help" href="http://www.w3.org/TR/workers/#apis-available-to-workers">
</head>
<body>
<script>
worker = startWorker("resources/worker-base64.js");
</script>
</body>
</html>
