third_party/WebKit/LayoutTests/fast/workers/worker-base64-expected.txt

[Worker] Tests that atob() / btoa() functions are exposed to workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/worker-base64.js
PASS [Worker] typeof atob === 'function' is true
PASS [Worker] typeof btoa === 'function' is true
PASS successfullyParsed is true

TEST COMPLETE
